requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
networkx>=3.2
pyyaml>=6.0
pandas>=2.1.0
